+++ KanbanApp.API/Entities/Scale.cs
@@ -3,8 +3,8 @@
     public class Scale
     {
         public int ScaleId { get; set; }
-        public float CalibrationFactor { get; set; }
-        public float InitialWeight{ get; set; }
+        public decimal CalibrationFactor { get; set; }
+        public decimal InitialWeight{ get; set; }
 
         // Nawigacja do Location (relacja jeden-do-jeden)
         public Location Location { get; set; }
